refactor(types): centralise shared domain types, drop dead ones
@meetpoint/types carried a parallel domain model (Meet, Participant, Place,
User, MeetFilters, ...) that nothing imported and that had drifted from the
Convex schema. Only Coordinates/TransportMode/Route/Isochrone were ever used.

- Trim location.ts to the types actually consumed; delete user.ts and the
  stale meet.ts model.
- Repurpose meet.ts to hold the genuinely shared domain types that were
  duplicated across the app: RsvpStatus (was redefined in 3 files) and
  SuggestedCity / SelectedCity (was in types/ai.ts + inlined in the meet page).
- types/ai.ts now re-exports from @meetpoint/types so @/types/ai stays stable;
  the meet page drops its local SelectedCity; the three RSVP files import the
  shared RsvpStatus.
