+++ packages/types/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./location";
 export * from "./meet";
-export * from "./user";
