+++ packages/types/src/meet.ts
@@ -1,60 +1,19 @@
-import type { Coordinates, Place, TransportMode } from "./location";
+import type { Coordinates } from "./location";
 
-export type MeetStatus = "draft" | "pending" | "confirmed" | "completed" | "cancelled";
+/** RSVP state of an event participant. Mirrors eventParticipants.rsvpStatus in the Convex schema. */
+export type RsvpStatus = "yes" | "no" | "maybe" | "pending";
 
-export type Participant = {
-  id: string;
-  userId?: string;
+/** A city proposed by the AI assistant. Mirrors meets.suggestedCities in the Convex schema. */
+export interface SuggestedCity {
   name: string;
-  location: Coordinates;
-  transportMode: TransportMode;
-  joinedAt: number;
-};
+  region: string;
+  coordinates: Coordinates;
+  reason: string;
+  matchScore: number;
+}
 
-export type LocationVote = {
-  placeId: string;
-  participantId: string;
-  vote: "up" | "down";
-  votedAt: number;
-};
-
-export type MeetFilters = {
-  maxTravelTimeMinutes?: number;
-  placeCategories?: string[];
-  priceLevel?: number[];
-  openNow?: boolean;
-};
-
-export type Meet = {
-  id: string;
+/** The city chosen as the effective destination. Mirrors meets.selectedCity in the Convex schema. */
+export interface SelectedCity {
   name: string;
-  description?: string;
-  creatorName: string;
-  participants: Participant[];
-  midpoint?: Coordinates;
-  suggestedPlaces: Place[];
-  votes: LocationVote[];
-  selectedPlace?: Place;
-  filters: MeetFilters;
-  status: MeetStatus;
-  scheduledFor?: number;
-  createdAt: number;
-  updatedAt: number;
-};
-
-export type CreateMeetInput = {
-  name: string;
-  description?: string;
-  creatorLocation: Coordinates;
-  creatorName: string;
-  transportMode: TransportMode;
-  filters?: MeetFilters;
-  scheduledFor?: number;
-};
-
-export type JoinMeetInput = {
-  meetId: string;
-  name: string;
-  location: Coordinates;
-  transportMode: TransportMode;
-};
+  coordinates: Coordinates;
+}
